Rework double parsing
