+++ src/base/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_library(
   base
   applicative.cpp
+  parserLibrary.cpp
   cbor.cpp
   enumOptionsBase.cpp
   geometry.cpp
